# gorm/models/project_li.go
package models

// Code generated by a tool. DO NOT EDIT.
// Additional custom methods can be implemented in a separate file like <this_file>_ex.go.
// https://sqldalmaker.sourceforge.net/

// a `project` list item
type ProjectLi struct {
	PId         int64  `json:"p_id" gorm:"column:p_id;primaryKey;autoIncrement"`
	PName       string `json:"p_name" gorm:"column:p_name;not null"`
	PTasksCount int64  `json:"p_tasks_count" gorm:"column:p_tasks_count;not null"`
}