# gorm/models/task_li.go
package models

// Code generated by a tool. DO NOT EDIT.
// Additional custom methods can be implemented in a separate file like <this_file>_ex.go.
// https://sqldalmaker.sourceforge.net/

// a `task` list item: no p_id, no t_comments
type TaskLi struct {
	TId       int64  `json:"t_id" gorm:"column:t_id;primaryKey;autoIncrement"`
	TPriority int64  `json:"t_priority" gorm:"column:t_priority;not null"`
	TDate     string `json:"t_date" gorm:"column:t_date;not null"`
	TSubject  string `json:"t_subject" gorm:"column:t_subject;not null"`
}