# no_orm/dbal/dto/task_li.go
package dto

// Code generated by a tool. DO NOT EDIT.
// Additional custom methods can be implemented in a separate file like <this_file>_ex.go.
// https://sqldalmaker.sourceforge.net/

// TaskLi list item (no p_id, no t_comments)
type TaskLi struct {
	TId       int64  `json:"t_id"`
	TPriority int64  `json:"t_priority"`
	TDate     string `json:"t_date"`
	TSubject  string `json:"t_subject"`
}